+++ ai-document-answer-retrieval-function/src/main/java/uk/gov/moj/cp/retrieval/service/SearchService.java
@@ -32,7 +32,7 @@ public class SearchService {
 
     public List<ChunkedEntry> searchDocumentsMatchingFilterCriteria(
             String userQuery,
-            List<Double> vectorizedUserQuery,
+            List<Float> vectorizedUserQuery,
             List<KeyValuePair> metadataFilters) {
 
         if (isNullOrEmpty(userQuery) || null == vectorizedUserQuery || vectorizedUserQuery.isEmpty() || null == metadataFilters || metadataFilters.isEmpty()) {
